ci: add a pipeline to validate publishable packages on pull requests
