+++ .github/workflows/validate-publishable-packages.yml
@@ -1,0 +1,22 @@
+name: Validate publishable packages
+
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+      - main
+    paths:
+      - 'packages/**'
+
+jobs:
+  validate-publishable-packages:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out repository code
+        uses: actions/checkout@v3
+
+      - name: install npm deps
+        run: npm ci
+
+      - name: publish shared package
+        run: npx ts-node tools/scripts/validate-publishable-packages.ts
